refactor: unify 404 page components
