+++ packages/core/client/src/pm/PluginManager.tsx
@@ -10,7 +10,7 @@
 export * from './PluginManagerLink';
 import { PageHeader } from '@ant-design/pro-layout';
 import { useDebounce } from 'ahooks';
-import { Button, Col, Divider, Input, List, Modal, Result, Row, Space, Spin, Table, Tabs } from 'antd';
+import { Button, Col, Divider, Input, List, Modal, Row, Space, Spin, Table, Tabs } from 'antd';
 import _ from 'lodash';
 import React, { useEffect, useMemo, useState } from 'react';
 import { useTranslation } from 'react-i18next';
@@ -19,6 +19,7 @@ import { useNavigate, useParams } from 'react-router-dom';
 import { css } from '@emotion/css';
 import { useACLRoleContext } from '../acl/ACLProvider';
 import { useAPIClient, useRequest } from '../api-client';
+import { AppNotFound } from '../common/AppNotFound';
 import { useDocumentTitle } from '../document-title';
 import { useToken } from '../style';
 import { PluginCard } from './PluginCard';
@@ -407,6 +408,6 @@ export const PluginManager = () => {
       </div>
     </div>
   ) : (
-    <Result status="404" title="404" subTitle="Sorry, the page you visited does not exist." />
+    <AppNotFound />
   );
 };
