+++ packages/core/client/src/pm/PluginSetting.tsx
@@ -9,11 +9,12 @@
 
 import { PageHeader } from '@ant-design/pro-layout';
 import { css } from '@emotion/css';
-import { Layout, Menu, Result } from 'antd';
+import { Layout, Menu } from 'antd';
 import _ from 'lodash';
 import React, { createContext, useCallback, useEffect, useMemo } from 'react';
 import { Navigate, Outlet, useLocation, useNavigate, useParams } from 'react-router-dom';
 import { ADMIN_SETTINGS_PATH, PluginSettingsPageType, useApp } from '../application';
+import { AppNotFound } from '../common/AppNotFound';
 import { useDocumentTitle } from '../document-title';
 import { useCompile } from '../schema-component';
 import { useStyles } from './style';
@@ -223,13 +224,7 @@ export const AdminSettingsLayout = () => {
               }
             />
           )}
-          <div className={styles.pageContent}>
-            {currentSetting ? (
-              <Outlet />
-            ) : (
-              <Result status="404" title="404" subTitle="Sorry, the page you visited does not exist." />
-            )}
-          </div>
+          <div className={styles.pageContent}>{currentSetting ? <Outlet /> : <AppNotFound />}</div>
         </Layout.Content>
       </Layout>
     </div>
